기능추가🔧 #42: 주변맛집 관련 api 연동 및 MVVM 리팩토링

+++ EatDa/Global/Network/Entities/RestaurantListModel.swift
@@ -12,7 +12,7 @@ struct RestaurantListModel: Decodable {
     let name: String?
     let explanation: String?
     let imgUrl: String?
-    let distance: Int?
+    let distance: Double?
     let hashTags: [String]
     let liked: Bool
 }
